nowPlaying + move features

+++ lib/cogs/Music.py
@@ -8,6 +8,7 @@ import logging
 from utils.Lavalink import LavalinkVoiceClient
 from utils.MusicFilters import FilterDropdownView, MusicFilters
 from utils.MusicQueue import QueueMenu, QueuePagination
+from StringProgressBar import progressBar
 import lavalink
 import re
 import functools
@@ -48,47 +49,45 @@ class Music(commands.Cog):
     def interaction_ensure_voice(f):
         @functools.wraps(f)
         async def callback(self, interaction: discord.Interaction, *args, **kwargs) -> None:
-            await self.ensure_voice(await discord.ext.commands.Context.from_interaction(interaction))
-            await f(self, interaction, *args, **kwargs)
+            if await self.ensure_voice(interaction):
+                await f(self, interaction, *args, **kwargs) 
         return callback
 
-    async def ensure_voice(self, ctx):
+    async def ensure_voice(self, interaction):
         """ This check ensures that the bot and command author are in the same voice channel. """
-        if ctx.guild is None:
+        if interaction.guild is None:
             raise commands.CommandInvokeError(
                 'This command can only be used inside a Discord server.')
         #  This is essentially the same as `@commands.guild_only()`
         #  except it saves us repeating ourselves (and also a few lines)
 
-        should_connect = ctx.command.name in ('play')
+        should_connect = interaction.command.name in ('play')
 
-        if not ctx.author.voice or not ctx.author.voice.channel:
-            # Our cog_command_error handler catches this and sends it to the voice channel.
-            # Exceptions allow us to "short-circuit" command invocation via checks so the
-            # execution state of the command goes no further.
-            raise commands.CommandInvokeError('Join a voice channel first.')
+        if not interaction.user.voice or not interaction.user.voice.channel:
+            await interaction.response.send_message(embed=self.bot.create_embed("MOCBOT MUSIC", f'Join a voice channel first.', None), ephemeral=True)
+            return False
 
-        v_client = ctx.voice_client
+        v_client = interaction.guild.voice_client
         if not v_client:
             if not should_connect:
-                raise commands.CommandInvokeError(
-                    "MOCBOT isn't connected to a voice channel.")
+                await interaction.response.send_message(embed=self.bot.create_embed("MOCBOT MUSIC", f'MOCBOT is not connected to a voice channel.', None), ephemeral=True)
+                return False
 
-            permissions = ctx.author.voice.channel.permissions_for(ctx.me)
+            permissions = interaction.user.voice.channel.permissions_for(interaction.guild.me)
 
             if not permissions.connect or not permissions.speak:  # Check user limit too?
-                raise commands.CommandInvokeError(
-                    'I need the `CONNECT` and `SPEAK` permissions.')
-
-            await ctx.author.voice.channel.connect(cls=LavalinkVoiceClient)
+                await interaction.response.send_message(embed=self.bot.create_embed("MOCBOT MUSIC", f'Please provide MOCBOT with the CONNECT and SPEAK permissions.', None), ephemeral=True)
+                return False
+            await interaction.user.voice.channel.connect(cls=LavalinkVoiceClient)
         else:
-            if v_client.channel.id != ctx.author.voice.channel.id:
-                raise commands.CommandInvokeError(
-                    'You need to be in my voice channel.')
+            if v_client.channel.id != interaction.user.voice.channel.id:
+                await interaction.response.send_message(embed=self.bot.create_embed("MOCBOT MUSIC", f'You need to be in my voice channel to execute that command.', None), ephemeral=True)
+                return False
 
-        player = self.bot.lavalink.player_manager.create(ctx.guild.id)
-        player.store('channel', ctx.channel.id)
+        player = self.bot.lavalink.player_manager.create(interaction.guild.id)
+        player.store('channel', interaction.channel.id)
         await player.set_volume(10)
+        return True
 
     async def track_hook(self, event):
         if isinstance(event, lavalink.events.QueueEndEvent):
@@ -99,10 +98,10 @@ class Music(commands.Cog):
             guild = self.bot.get_guild(guild_id)
             if guild_id in self.players:
                 channel = guild.get_channel(self.players[guild_id]["CHANNEL"])
-                message = await channel.fetch_message(self.players[guild_id]["MESSAGE_ID"])
-
+                message = await self.retrieve_now_playing(channel, guild)
                 await guild.voice_client.disconnect(force=True)
-                await message.delete()
+                if message is not None:
+                    await message.delete()
                 del self.players[guild_id]
 
     async def next_playing(self, event):
@@ -143,12 +142,20 @@ class Music(commands.Cog):
 
     async def sendNewNowPlaying(self, guild, player):
         channel = guild.get_channel(self.players[guild.id]["CHANNEL"])
-        message = await channel.fetch_message(self.players[guild.id]["MESSAGE_ID"])
+        message = await self.retrieve_now_playing(channel, guild)
         if not self.players[guild.id]["FIRST"]:
-            await message.delete()
+            if message is not None:
+                await message.delete()
             message = await channel.send(embed=await self.generateNowPlayingEmbed(guild, player))
-        self.players[guild.id] = {
-            "CHANNEL": channel.id, "MESSAGE_ID": message.id, "FIRST": False}
+        self.players[guild.id] = {"CHANNEL": channel.id, "MESSAGE_ID": message.id, "FIRST": False}
+
+    async def retrieve_now_playing(self, channel, guild):
+        try:
+            message = await channel.fetch_message(self.players[guild.id]["MESSAGE_ID"])
+        except discord.errors.NotFound:
+            return None
+        else:
+            return message
 
     async def progress_update(self, event):
         if isinstance(event, lavalink.events.PlayerUpdateEvent):
@@ -229,7 +236,8 @@ class Music(commands.Cog):
             embed.set_image(url=await self.getMediaThumbnail(player.queue[-1].source_name, player.queue[-1].identifier))
             embed.add_field(name="POSITION", value=len(
                 player.queue), inline=True)
-            embed.add_field(name="QUEUE TIME", value=await self.formatDuration(reduce(lambda a, b: a + b, [song.duration if not song.stream else 0 for song in player.queue])), inline=True)
+            duration = reduce(lambda a, b: a + b, [song.duration if not song.stream else 0 for song in player.queue])
+            embed.add_field(name="QUEUE TIME", value=await self.formatDuration(duration) if (duration < 86400000) else '>24h', inline=True)
             embed.set_footer(text=f"Requested by {interaction.user}")
             await interaction.followup.send(embed=embed)
             await self.delay_delete(interaction, Music.MESSAGE_ALIVE_TIME)
@@ -262,6 +270,7 @@ class Music(commands.Cog):
     @app_commands.describe(
         position="The queue item number to skip to."
     )
+    @interaction_ensure_voice
     async def skip(self, interaction: discord.Interaction, position: typing.Optional[int] = 1):
         player = self.bot.lavalink.player_manager.get(interaction.guild.id)
         if player is None or player.current is None:
@@ -285,6 +294,7 @@ class Music(commands.Cog):
     @app_commands.describe(
         time="The time in seconds to seek to."
     )
+    @interaction_ensure_voice
     async def seek(self, interaction: discord.Interaction, time: int):
         player = self.bot.lavalink.player_manager.get(interaction.guild.id)
         if player is None or player.current is None:
@@ -301,6 +311,7 @@ class Music(commands.Cog):
         await self.delay_delete(interaction, 5)
 
     @app_commands.command(name="loop", description="Loop the current media or queue.")
+    @interaction_ensure_voice
     async def loop(self, interaction: discord.Interaction, type: Literal["Off", "Song", "Queue"]):
         player = self.bot.lavalink.player_manager.get(interaction.guild.id)
         if player is None or player.current is None:
@@ -320,20 +331,10 @@ class Music(commands.Cog):
         await self.delay_delete(interaction, 5)
 
     @app_commands.command(name="disconnect", description="Disconnects the bot from voice.")
+    @interaction_ensure_voice
     async def disconnect(self, interaction: discord.Interaction):
         """ Disconnects the player from the voice channel and clears its queue. """
         player = self.bot.lavalink.player_manager.get(interaction.guild.id)
-
-        if not interaction.guild.voice_client:
-            # We can't disconnect, if we're not connected.
-            await interaction.response.send_message(embed=self.bot.create_embed("MOCBOT MUSIC", f"MOCBOT isn't connected to a voice channel.", None))
-            return await self.delay_delete(interaction, Music.MESSAGE_ALIVE_TIME)
-
-        if not interaction.user.voice or (player.is_connected and interaction.user.voice.channel.id != int(player.channel_id)):
-            # Abuse prevention. Users not in voice channels, or not in the same voice channel as the bot
-            # may not disconnect the bot.
-            await interaction.response.send_message(embed=self.bot.create_embed("MOCBOT MUSIC", f"You must be in the same channel as MOCBOT to use this command.", None))
-            return await self.delay_delete(interaction, Music.MESSAGE_ALIVE_TIME)
 
         # Clear the queue to ensure old tracks don't start playing
         # when someone else queues something.
@@ -344,8 +345,9 @@ class Music(commands.Cog):
         if interaction.guild.id in self.players:
             channel = interaction.guild.get_channel(
                 self.players[interaction.guild.id]["CHANNEL"])
-            message = await channel.fetch_message(self.players[interaction.guild.id]["MESSAGE_ID"])
-            await message.delete()
+            message = await self.retrieve_now_playing(channel, interaction.guild)
+            if message is not None:
+                await message.delete()
             del self.players[interaction.guild.id]
 
         # Disconnect from the voice channel.
@@ -354,6 +356,7 @@ class Music(commands.Cog):
         await self.delay_delete(interaction, Music.MESSAGE_ALIVE_TIME)
 
     @app_commands.command(name="stop", description="Stops any media that is playing.")
+    @interaction_ensure_voice
     async def stop(self, interaction: discord.Interaction):
         """ Stops the player. """
         player = self.bot.lavalink.player_manager.get(interaction.guild.id)
@@ -376,14 +379,16 @@ class Music(commands.Cog):
         if interaction.guild.id in self.players:
             channel = interaction.guild.get_channel(
                 self.players[interaction.guild.id]["CHANNEL"])
-            message = await channel.fetch_message(self.players[interaction.guild.id]["MESSAGE_ID"])
-            await message.delete()
+            message = await self.retrieve_now_playing(channel, interaction.guild)
+            if message is not None:
+                await message.delete()
             del self.players[interaction.guild.id]
 
         await interaction.response.send_message(embed=self.bot.create_embed("MOCBOT MUSIC", f"MOCBOT has been stopped.", None))
         await self.delay_delete(interaction, Music.MESSAGE_ALIVE_TIME)
 
     @app_commands.command(name="filters", description="Toggles audio filters")
+    @interaction_ensure_voice
     async def filters(self,  interaction: discord.Interaction):
         player = self.bot.lavalink.player_manager.get(interaction.guild.id)
         if player is None or player.current is None:
@@ -393,6 +398,7 @@ class Music(commands.Cog):
 
     # Written by Sam https://github.com/sam1357
     @app_commands.command(name="pause", description="Pauses the music")
+    @interaction_ensure_voice
     async def pause(self, interaction: discord.Interaction):
         player = self.bot.lavalink.player_manager.get(interaction.guild.id)
 
@@ -414,6 +420,7 @@ class Music(commands.Cog):
 
     # Written by Sam https://github.com/sam1357
     @app_commands.command(name="resume", description="Resumes the music")
+    @interaction_ensure_voice
     async def resume(self, interaction: discord.Interaction):
         player = self.bot.lavalink.player_manager.get(interaction.guild.id)
 
@@ -434,6 +441,7 @@ class Music(commands.Cog):
         await self.delay_delete(interaction, Music.MESSAGE_ALIVE_TIME)
 
     @app_commands.command(name="shuffle", description="Shuffles the queue")
+    @interaction_ensure_voice
     async def shuffle(self, interaction: discord.Interaction):
         player = self.bot.lavalink.player_manager.get(interaction.guild.id)
 
@@ -451,6 +459,7 @@ class Music(commands.Cog):
         start="The track number to remove from",
         end="The track number to remove to (optional)"
     )
+    @interaction_ensure_voice
     async def remove(self, interaction: discord.Interaction, start: int, end: typing.Optional[int]):
         player = self.bot.lavalink.player_manager.get(interaction.guild.id)
         removedTrack = None
@@ -471,9 +480,54 @@ class Music(commands.Cog):
                 await interaction.response.send_message(embed=self.bot.create_embed("MOCBOT MUSIC", f"Please enter a valid range to remove. There {'is' if len(player.queue) == 1 else 'are'} **{len(player.queue)}** track{'' if len(player.queue) == 1 else 's'} in the queue.", None))
                 return await self.delay_delete(interaction, Music.MESSAGE_ALIVE_TIME)
             player.queue = player.queue[:start - 1] + player.queue[end:]
-            await interaction.response.send_message(embed=self.bot.create_embed("MOCBOT MUSIC", f"Successfully removed {end - start + 1} track{'' if end - start + 1 == 1 else 's'} from the queue.", None))
+            await interaction.response.send_message(embed=self.bot.create_embed("MOCBOT MUSIC", f"Successfully removed **{end - start + 1} track{'' if end - start + 1 == 1 else 's'}** from the queue.", None))
             return await self.delay_delete(interaction, 5)
 
+    @app_commands.command(name="move", description="Moves the given track to another position in the queue.")
+    @app_commands.describe(
+        source="The track number to move",
+        destination="The position in the queue to move to"
+    )
+    @interaction_ensure_voice
+    async def move(self, interaction: discord.Interaction, source: int, destination: int):
+        player = self.bot.lavalink.player_manager.get(interaction.guild.id)
+
+        if player is None or player.current is None:
+            await interaction.response.send_message(embed=self.bot.create_embed("MOCBOT MUSIC", f"The move command requires media to be playing first.", None))
+            return await self.delay_delete(interaction, Music.MESSAGE_ALIVE_TIME)
+        if len(player.queue) < 2:
+            await interaction.response.send_message(embed=self.bot.create_embed("MOCBOT MUSIC", f"There are less than two songs in the queue.", None))
+            return await self.delay_delete(interaction, Music.MESSAGE_ALIVE_TIME)
+        if source == destination: 
+            await interaction.response.send_message(embed=self.bot.create_embed("MOCBOT MUSIC", f"You can only move songs to a different position in the queue.", None))
+            return await self.delay_delete(interaction, Music.MESSAGE_ALIVE_TIME)
+        if source < 1 or source > len(player.queue) or destination < 1 or destination > len(player.queue):
+            await interaction.response.send_message(embed=self.bot.create_embed("MOCBOT MUSIC", f"Please enter valid positions in the queue. There {'is' if len(player.queue) == 1 else 'are'} **{len(player.queue)}** track{'' if len(player.queue) == 1 else 's'} in the queue.", None))
+            return await self.delay_delete(interaction, Music.MESSAGE_ALIVE_TIME)
+
+        player.queue.insert(destination - 1, track:=player.queue.pop(source - 1))
+        await interaction.response.send_message(embed=self.bot.create_embed("MOCBOT MUSIC", f"Successfully moved track [{track.title}]({track.uri}) to position **{destination}** in the queue.", None))
+        return await self.delay_delete(interaction, Music.MESSAGE_ALIVE_TIME)
+
+    @app_commands.command(name="nowplaying", description="Sends a message regarding the currently playing song and its progress")
+    async def now_playing(self, interaction: discord.Interaction):
+        player = self.bot.lavalink.player_manager.get(interaction.guild.id)
+
+        if player is None or player.current is None:
+            await interaction.response.send_message(embed=self.bot.create_embed("MOCBOT MUSIC", f"The now playing command requires media to be playing first.", None))
+            return await self.delay_delete(interaction, Music.MESSAGE_ALIVE_TIME)
+        progress_bar = progressBar.splitBar(player.current.duration, int(player.position), size=15)
+
+        embed = self.bot.create_embed(
+            "MOCBOT MUSIC", f"> NOW PLAYING: [{player.current.title}]({player.current.uri})", None)
+        embed.add_field(name="Requested By", value=f'<@{player.current.requester}>', inline=True)
+        embed.add_field(name="Uploader",
+                        value=player.current.author, inline=True)
+        embed.add_field(name="Progress",
+                        value=f'{datetime.datetime.utcfromtimestamp(int(player.position) / 1000).strftime("%H:%M:%S")} {progress_bar[0]} {"🔴 LIVE" if player.current.stream else datetime.datetime.utcfromtimestamp(int(player.current.duration) / 1000).strftime("%H:%M:%S")}', inline=False)
+        embed.set_thumbnail(url=await self.getMediaThumbnail(player.current.source_name, player.current.identifier))
+        await interaction.response.send_message(embed=embed)
+        return await self.delay_delete(interaction, Music.MESSAGE_ALIVE_TIME * 2)
 
 async def setup(bot):
     await bot.add_cog(Music(bot))
